### 2023-02-21 #5178 @editorialbot
[PRE REVIEW]: ViMag: A Visual Vibration Analysis Toolbox
**Submitting author:** @RLado (<a href="http://orcid.org/0000-0002-6421-7351">Ricard Lado Roigé</a>)
**Repository:** https://github.com/RLado/ViMag
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.1.6
