### 2023-02-21 #5178 @RLado
Hello @arfon.

The submodule [STB-VMM [1]](https://github.com/RLado/STB-VMM) corresponds to a [previous submission](https://arxiv.org/abs/2302.10001) to Knowledge-Based Systems (under review) and is used as the motion magnification backend for ViMag.

The other included Python submodules [[2]](https://github.com/RLado/TempSlice/tree/db3b2d9651f9876161c520a6e0172e73e38b9216)[[3]](https://github.com/RLado/Tiled-STB-VMM/tree/fd51d61f74194a249292134403fc100505ca8deb) are part of the core functionality of ViMag and do not correspond to any previous publications. I have used these submodules as standalone CLI tools to verify our damage detection methodology, for this reason they are presented in this format. The javascript code corresponds to the electron app that combines these tools into a user-friendly, multiplatform and practical interface.

Sorry for any confusion caused. And thank you for your time.
