### 2023-03-04 #5178 @arfon
Thanks @RLado – I'm going to ask for the input of the wider JOSS editorial team here on whether we think this submission is in scope for JOSS.

Note to JOSS editors: These are the two submodules being references above:

```
cloc TempSlice/
       8 text files.
       8 unique files.
       1 file ignored.

Language                     files          blank        comment           code
Python                           6            107             72            453
Markdown                         1             33              0            139
Text                             1              0              0             12
SUM:                             8            140             72            604
```

```
cloc Tiled-STB-VMM/
       6 text files.
       4 unique files.
       3 files ignored.

Language                     files          blank        comment           code
Python                           1             45             48            199
Markdown                         2             16              0             62
Text                             1              0              0              5
SUM:                             4             61             48            266
```
