+++ README.md
@@ -1,5 +1,6 @@
 
 # <img title="Logo" alt="Logo: A tuning fork with an eye" src="./img/icon.svg" width="40px"></img>    ViMag: *A Visual Vibration Toolbox*
+[![status](https://joss.theoj.org/papers/a0149e43cf19bbafe1cb9eecdcde6189/status.svg)](https://joss.theoj.org/papers/a0149e43cf19bbafe1cb9eecdcde6189)
 
 Vision-based damage detection techinques can reduce sensor deployment costs while providing accurate, useful, and full-field readings of structural behaviour. Our work presents a video processing methodology based on [STB-VMM](https://github.com/RLado/STB-VMM) implemented in a software toolbox that allows the processing of video data to obtain vibrational signatures of complex structures. Our [tests](https://authors.elsevier.com/c/1gATkxsQaIVc~) have demonstrated that this technique enables the identification of very light structural damage in a controlled lab environament. This vision-based technique may not be as precise as other contact-based or laser methods, but on the other hand, it offers an easy to use, efective, full-field, tool for structural health monitoring at a fraction of the cost.
 
