### 2023-05-08 67df187
Changed motion magnification backend for testing purposes
