### 2023-05-18 #5178 @Kevin-Mattheus-Moerman
@iamtsac @ZHANGKEON @capepoint  would you be interested in reviewing a paper for the [Journal of Open Source Software (JOSS)](https://joss.theoj.org/) entitled: "ViMag: A Visual Vibration Analysis Toolbox"?

JOSS reviews focus on the [software](https://github.com/RLado/ViMag) as well as a [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05178/joss.05178/10.21105.joss.05178.pdf)

JOSS reviews take place on GitHub and reviewers are guided by a set of check boxes. Here is an example of an ongoing review: https://github.com/openjournals/joss-reviews/issues/5278.

You can let me know here if you are interested. Let me know if you have any question.
