### 2023-05-18 #5178 @josepfontm
@Kevin-Mattheus-Moerman Thanks for considering me as a possible reviewer. Unfortunately, I collaborate with @RLado frequently, as you correctly assumed I am not a viable reviewer.
Nonetheless, I have some suggestions which might interest you. All the reviewers I propose are researchers working in the same field as us: Structural Health Monitoring/Vibration Analysis and have active communities on Github.

I would suggest contacting the people behind the PyOMA project.
https://github.com/dagghe/PyOMA

In addition, the research group from Professor Janko Slavič has wide experience in open-source projects and, more specifically, in vibration analysis. I attach their website:
http://lab.fs.uni-lj.si/ladisk/?

Best,
