### 2023-06-04 #5491 @jankoslavic
Dear authors!

Congratulations on the effort! While the comments below might sound critical, I hope they help to increase the impact of your work.
I will start with the pdf article: the article should have referenced the original prior knowledge; as i understand, this work relates on Swin Transformer Based Video Motion Magnification; however, prior research on Motion Magnification (especially by Freeman, W.) should be mentioned.

Further:
1. ViMag is built on top of https://github.com/RLado/STB-VMM/. Would it be possible that STB-VMM would be prepared as a separate package which would be included here as a requirement? Similar TempSlice, Tiled-SBT-VMM; currently in linked repositories look hard to maintain and the code is not clear / well decoupled. Separate packages should also have a clear documentation and testing procedures.
2. I have downloaded and run the release files without a problem. I would suggest that you add example files to the repository. The link to "tests": https://authors.elsevier.com/c/1gATkxsQaIVc in the readme is not working.
3. Try to make the User interface of the program more intuitive: e.g. use icons that explain a particular task better (currently the icons are generic); try to improve the flow of the program (e.g.: define >> identify motion >> analyze)
4. The application is huge (3.3GB on Win). Why not use a simple Python UI?
