### 2023-06-10 54cd74f
Added a paragraph briefly introducing relevant VMM works
