+++ paper/paper.bib
@@ -158,3 +158,42 @@
 	isbn="978-3-030-59716-0",
 	doi={10.48550/arXiv.2009.07432}
 }
+
+@inproceedings{liu_motion_2005,
+	author = {Liu, Ce and Torralba, Antonio and Freeman, William T. and Durand, Fr\'{e}do and Adelson, Edward H.},
+	title = {Motion Magnification},
+	year = {2005},
+	isbn = {9781450378253},
+	//publisher = {Association for Computing Machinery},
+	//address = {New York, NY, USA},
+
+	doi = {10.1145/1186822.1073223},
+	abstract = {We present motion magnification, a technique that acts like a microscope for visual motion. It can amplify subtle motions in a video sequence, allowing for visualization of deformations that would otherwise be invisible. To achieve motion magnification, we need to accurately measure visual motions, and group the pixels to be modified. After an initial image registration step, we measure motion by a robust analysis of feature point trajectories, and segment pixels based on similarity of position, color, and motion. A novel measure of motion similarity groups even very small motions according to correlation over time, which often relates to physical cause. An outlier mask marks observations not explained by our layered motion model, and those pixels are simply reproduced on the output from the original registered observations.The motion of any selected layer may be magnified by a user-specified amount; texture synthesis fills-in unseen "holes" revealed by the amplified motions. The resulting motion-magnified images can reveal or emphasize small motions in the original sequence, as we demonstrate with deformations in load-bearing structures, subtle motions or balancing corrections of people, and "rigid" structures bending under hand pressure.},
+	booktitle = {ACM SIGGRAPH 2005 Papers},
+	pages = {519–526},
+	numpages = {8},
+	keywords = {video processing, video-based rendering, computer vision, motion processing},
+	location = {Los Angeles, California},
+	//series = {SIGGRAPH '05}
+}
+
+@article{wu_eulerian_2012,
+	abstract = {Using the Eulerian Video Magnification method, the flow of blood as it fills the face is able to be visualize and the resulting signal is amplified to reveal hidden information. Our goal is to reveal temporal variations in videos that are difficult or impossible to see with the naked eye and display them in an indicative manner. Our method, which we call Eulerian Video Magnification, takes a standard video sequence as input, and applies spatial decomposition, followed by temporal filtering to the frames. The resulting signal is then amplified to reveal hidden information. Using our method, we are able to visualize the flow of blood as it fills the face and also to amplify and reveal small motions. Our technique can run in real time to show phenomena occurring at the temporal frequencies selected by the user.},
+	author = {Wu, Hao-Yu and Rubinstein, Michael and Shih, Eugene and Guttag, J. and Durand, F. and Freeman, W.},
+	doi = {10.1145/2185520.2185561},
+	journal = {ACM Trans. Graph.},
+	title = {Eulerian video magnification for revealing subtle changes in the world},
+	year = {2012}
+}
+
+@inproceedings{wadhwa_riesz_2014,
+	abstract = {We present a new compact image pyramid representation, the Riesz pyramid, that can be used for real-time phase-based motion magnification. Our new representation is less overcomplete than even the smallest two orientation, octave-bandwidth complex steerable pyramid, and can be implemented using compact, efficient linear filters in the spatial domain. Motion-magnified videos produced with this new representation are of comparable quality to those produced with the complex steerable pyramid. When used with phase-based video magnification, the Riesz pyramid phase-shifts image features along only their dominant orientation rather than every orientation like the complex steerable pyramid.},
+	author = {Wadhwa, Neal and Rubinstein, Michael and Durand, Fredo and Freeman, William},
[…]
