+++ paper/paper.bib
@@ -164,8 +164,6 @@
 	title = {Motion Magnification},
 	year = {2005},
 	isbn = {9781450378253},
-	//publisher = {Association for Computing Machinery},
-	//address = {New York, NY, USA},
 
 	doi = {10.1145/1186822.1073223},
 	abstract = {We present motion magnification, a technique that acts like a microscope for visual motion. It can amplify subtle motions in a video sequence, allowing for visualization of deformations that would otherwise be invisible. To achieve motion magnification, we need to accurately measure visual motions, and group the pixels to be modified. After an initial image registration step, we measure motion by a robust analysis of feature point trajectories, and segment pixels based on similarity of position, color, and motion. A novel measure of motion similarity groups even very small motions according to correlation over time, which often relates to physical cause. An outlier mask marks observations not explained by our layered motion model, and those pixels are simply reproduced on the output from the original registered observations.The motion of any selected layer may be magnified by a user-specified amount; texture synthesis fills-in unseen "holes" revealed by the amplified motions. The resulting motion-magnified images can reveal or emphasize small motions in the original sequence, as we demonstrate with deformations in load-bearing structures, subtle motions or balancing corrections of people, and "rigid" structures bending under hand pressure.},
@@ -174,7 +172,6 @@
 	numpages = {8},
 	keywords = {video processing, video-based rendering, computer vision, motion processing},
 	location = {Los Angeles, California},
-	//series = {SIGGRAPH '05}
 }
 
 @article{wu_eulerian_2012,
