### 2023-06-10 60d49a1
Updated STB-VMM citation
