### 2023-06-10 2e2d65b
Added some more context to VMM and moved fig 1 up
