+++ paper/paper.md
@@ -36,11 +36,9 @@ This work is based on the method developed by @LADOROIGE2022112218 for vibration
 
 Other researchers have used similar techniques for vibration testing [@MOLINAVIEDMA2018245; @EITNER2021106995]. However, to the authors' knowledge, non have released a software tool to go along with their publications. ViMag offers a simple interface to replicate some of these experiments using state-of-the-art learning-based video motion magnification.
 
-<!-- Motion magnification is a video processing technique that consists on the transformation of input frames to exaggerate motion. The goal of these algorithms is to amplify subtle motions in a video sequence, allowing the visualization of vibrations and deformations that would otherwise be invisible. -->
-
-Video motion magnification was first developed by @liu_motion_2005 and opened a new range of possibilities for research, however, this first approach produced numerous visual artifacts on top of being computationally expensive. Years later,further developments by @wu_eulerian_2012 introduced a novel Eulerian approach to magnification that produced much cleaner results with less computational cost, paving the way for newer and more refined algorithms that produced increasingly better results such as [@wadhwa_riesz_2014], [@oh_learning-based_2018] or [@lado2022_STB-VMM].
-
 ![Video sequence to signal pipeline \label{fig:pipeline}](pipeline_chart.png){ width=90% }
+
+Motion magnification is a video processing technique that consists on the transformation of input frames to exaggerate motion. The goal of these algorithms is to amplify subtle motions in a video sequence, allowing the visualization of vibrations and deformations that would otherwise be invisible. Video motion magnification was first developed by @liu_motion_2005 and opened a new range of possibilities for research, however, this first approach produced numerous visual artifacts on top of being computationally expensive. Years later,further developments by @wu_eulerian_2012 introduced a novel Eulerian approach to magnification that produced much cleaner results with less computational cost, paving the way for newer and more refined algorithms that produced increasingly better results such as [@wadhwa_riesz_2014], [@oh_learning-based_2018] or [@lado2022_STB-VMM].
 
 # Statement of need
 
