### 2023-06-10 514e2a5
Updated README page - Removed use/build instrucctions from the README page. In their place now there is a link to the project's wiki - "tests" link no longer points to the custom elsevier share link (now expired)
