+++ README.md
@@ -1,74 +1,25 @@
 
 # <img title="Logo" alt="Logo: A tuning fork with an eye" src="./img/icon.svg" width="40px"></img>    ViMag: *A Visual Vibration Toolbox*
-[![status](https://joss.theoj.org/papers/a0149e43cf19bbafe1cb9eecdcde6189/status.svg)](https://joss.theoj.org/papers/a0149e43cf19bbafe1cb9eecdcde6189)
 
-Vision-based damage detection techinques can reduce sensor deployment costs while providing accurate, useful, and full-field readings of structural behaviour. Our work presents a video processing methodology based on [STB-VMM](https://github.com/RLado/STB-VMM) implemented in a software toolbox that allows the processing of video data to obtain vibrational signatures of complex structures. Our [tests](https://authors.elsevier.com/c/1gATkxsQaIVc~) have demonstrated that this technique enables the identification of very light structural damage in a controlled lab environament. This vision-based technique may not be as precise as other contact-based or laser methods, but on the other hand, it offers an easy to use, efective, full-field, tool for structural health monitoring at a fraction of the cost.
+<a href="https://joss.theoj.org/papers/a0149e43cf19bbafe1cb9eecdcde6189"><img src="https://joss.theoj.org/papers/a0149e43cf19bbafe1cb9eecdcde6189/status.svg"></a>
 
-## Initial setup
+Vision-based damage detection techinques can reduce sensor deployment costs while providing accurate, useful, and full-field readings of structural behaviour. Our work presents a video processing methodology based on [STB-VMM](https://github.com/RLado/STB-VMM) implemented in a software toolbox that allows the processing of video data to obtain vibrational signatures of complex structures. Our [tests](https://doi.org/10.1016/j.measurement.2022.112218) have demonstrated that this technique enables the identification of very light structural damage in a controlled lab environment. This vision-based technique may not be as precise as other contact-based or laser methods, but on the other hand, it offers an easy to use, effective, full-field, tool for structural health monitoring at a fraction of the cost.
 
-### Linux 
-- Run the packaged executable
+## Features
+- Apply state-of-the-art video motion magnification
+- Select multiple video areas for displacement monitoring
+- Obtain discrete displacement-over-time signals of selected areas
+- Compute the frequency response of displacement-over-time signals
+- Export the results in .csv for further processing
 
-### Windows
-- Install Microsoft Visual C++ Redistributable from [here](https://aka.ms/vs/16/release/vc_redist.x64.exe) (if not installed already)
-- Run the packaged executable
+<hr style="margin-top: 20px; margin-bottom: 20px;">
 
+**Download the latest [release](https://github.com/RLado/ViMag/releases)**
 
-## Use instrucctions
+Use instructions & documentation can be found in the [wiki](https://github.com/RLado/ViMag/wiki)
 
-To start working with ViMag import a video file using the *File>Import video* menu or click the import icon in the left:
+<hr style="margin-top: 20px; margin-bottom: 20px;">
 
-<img src="https://user-images.githubusercontent.com/25719985/203530586-c6216f03-5cdb-4226-bdd1-7cd3635ce857.png" width="600px"></img>
-
-A dialog will open and let you select your desired video file(s). It is advised that input videos be recorded using a tripod and be no longer than 10 seconds approx.
-
-Next, add slices to the areas in which you want to monitor vibration. This areas must have a contrasting color from the background. To add slices switch *Slice Mode* on by clicking the button. Then click once on the video to define the first point of the slice and again to define a second. Now your first slice should appear on screen. 
-
[…]
