### 2023-06-11 #5491 @RLado
Dear @jankoslavic,

Thank you for your time and your comments. Your criticism is well received and I would like to add some context on why I made some of the decisions regarding ViMag's software stack.

I conceived ViMag to be as portable as possible so I can grab any computer from the lab, connect a thumbdrive with the program and get to work without worrying about dependencies. Also, I wanted ViMag to be cross-platform because I use Linux but our university mostly runs on Windows. Finally, I thought that using web technologies for coding the UI would be easier, feel more modern, and could be expanded by students with web development knowledge. These requirements finally let me to choosing Electron, a very popular framework to build cross-platform desktop apps using web technologies.

Electron is not particularly lightweight as it is based on chromium but has great tools for packaging to multiple platforms, which is great for a solo developer. However, that does not excuse a 3.3GB package. Most of that space is taken by the python interpreter and the required dependencies bundled, pytorch can take up to 1.1GB alone. Indeed I am bundling cpython in the package, I know that pyinstaller and nuitka exist but they did not work properly or generated binaries much larger than cpython + dependencies. I guess the alternative would be porting all python code to nodejs, which should be possible, but not using pytorch. In hindsight, my choice to make a portable application using python components resulted in a not very elegant stack. Nevertheless, after using ViMag for a while I am happy with the results and is definitely much better that selecting pixel coordinates on the CLI.

Other comments:
 - In regards to the possibility of bundling STB-VMM and other python code into installable packages it would be possible using wheels. However, I am not sure this would remove complexity from the current stack.

 - I have changed the `tests` link which was provided by elsevier to authors for sharing their papers for free (for 50 days, then expires). Now the link points to the doi of the article directly. Test files are linked in the [wiki](https://github.com/RLado/ViMag/wiki) under the testing section.

 - I have removed the 'use instructions' section from the README file. It is better to direct users to the wiki so only one file needs to be updated in case the UI is updated.

 - I agree that the user interface needs a redesign, but I would like to have some more feedback before I start changing things. Proper UI design is complicated, time consuming and requires lots of testing.

 - In regards to the pdf article, I have added a paragraph to the "Related work" section briefly talking about motion magnification where I mention relevant prior research.

---
