### 2023-06-19 #5491 @iamtsac
Great work on this handy application. In general, such software can be rather helpful in fields such as SHM, where different types of sensors are employed to monitor a structure’s vibration, which, however, are accompanied by the disadvantages of noise readings and the hustle of calibration.

Some comments I have to add are:

- Such an application has a multi-disciplinary nature, and different scientific fields can utilize it. Thus the UI should be more informative. An example would be a progress bar to show the remaining computation time. Also, the cog icon is commonly used for the settings button, so it is not intuitive that starts the calculations. The magnification settings should be in the main panel and easily accessible by the user rather than in the preferences dialog.

- During the installation, I encountered an error with the torch library related to the cython version on my end. However, the software did not provide any error messages, so I had to figure it out myself. Thus, I believe some additional error checking could be beneficial.

- Finally, the computation time is a bit slow, which is reasonable given the used model. However, during the computation, I noticed that the GPU is not utilized hence exploiting the GPU, when available, is essential.
