### 2023-07-05 #5491 @iamtsac
@RLado

Sadly, I could not reproduce the installation problem to provide additional details. However, the installation wiki page is thorough and on-point. Thus I do not have any changes to suggest. Despite all that, this minor error pointed me to check on the error handling in your package as a whole, and I think it is more important to focus on that, which will result in a more user-friendly application.

I am glad my suggestions are well received and will probably lead to a well-polished application. On my end, I do not have any other recommendations.
