+++ main.js
@@ -1,299 +1,30 @@
-// Import
-const fs = require('fs');
-const process = require('process');
-const path = require('path');
-const { app, BrowserWindow, Menu, ipcMain } = require('electron');
-const Main = require('electron/main');
-let { PythonShell } = require('python-shell');
+const { app, screen, BrowserWindow } = require('electron')
 
-
-// Setup python environament
-console.log('Setting up python environament ...');
-let pyvenvContent;
-if (process.platform === 'win32') {
-    pyvenvContent = `home = ${path.join(process.cwd(), 'python/interpreter/PCbuild/amd64')}
-include-system-site-packages = false
-version = 3.10.6
-`;
-}
-else {
-    pyvenvContent = `home = ${path.join(process.cwd(), 'python/interpreter')}
-include-system-site-packages = false
-version = 3.10.6
-`;
-}
-fs.writeFile(path.join(process.cwd(), 'python/interpreter/ViMag_venv/pyvenv.cfg'), pyvenvContent, err => {
-    if (err) {
-        console.error(err);
-    }
-    // file written successfully
-});
-console.log('Done');
-
-// Add application icon
-app.getFileIcon('./img/icon.ico');
-
-// Define main window
 const createWindow = () => {
+    const primaryDisplay = screen.getPrimaryDisplay()
+    const { width, height } = primaryDisplay.size
     const win = new BrowserWindow({
-        title: 'ViMag',
-        width: 1200,
-        height: 720,
-        icon: './img/icon.ico',
-        fullscreenable: false,
-        //frame: False,
-        webPreferences: {
-            nodeIntegration: true, // These arguments are now required since they are disabled by default for security reasons
-            contextIsolation: false,
-            devTools: !app.isPackaged,
-        },
-    });
-    win.toggleDevTools(); //For debug
-    win.loadFile('./html/index.html');
-
-    const menuTemplate = [
-        {
-            label: 'File',
-            submenu: [
-                {
-                    label: 'New project',
-                    accelerator: 'CmdOrCtrl+N',
-                    click() {
-                        win.webContents.send('newPrj', true);
-                    }
-                },
-                {
-                    label: 'Load project',
-                    accelerator: 'CmdOrCtrl+O',
-                    click() {
-                        loadProject(win);
-                    }
-                },
-                {
-                    label: 'Save',
-                    accelerator: 'CmdOrCtrl+S',
-                    click() {
-                        win.webContents.send('save', null);
-                    }
-                },
-                {
-                    label: 'Save As',
-                    accelerator: 'CmdOrCtrl+Shift+S',
-                    click() {
-                        saveProject(win);
-                    }
-                },
-                {
-                    type: 'separator'
-                },
-                {
-                    label: 'Import video',
-                    accelerator: 'CmdOrCtrl+I',
-                    click() {
-                        importVid(win);
-                    },
-                },
-                {
-                    type: 'separator'
-                },
-                {
-                    role: 'Quit'
-                }
-            ]
-        },
-
-        {
-            label: 'Edit',
-            submenu: [
-                {
-                    label: 'Preferences',
-                    click() {
-                        const aboutWin = new BrowserWindow({
-                            title: 'Preferences',
-                            width: 600,
-                            height: 800,
-                            icon: './img/icon.ico',
-                            resizable: false,
-                            fullscreenable: false,
-                            webPreferences: {
-                                nodeIntegration: true, // These arguments are now required since they are disabled by default for security reasons
[…]
