+++ img/svg2ico.sh
@@ -1,1 +1,0 @@
-convert -density 300 -define icon:auto-resize=256,128,96,64,48,32,16 -background none icon.svg icon.ico
