+++ js/renderer.js
@@ -1,1813 +1,0 @@
-// Import ----------------------------------------------------------------------
-const { ipcRenderer } = require('electron');
-let Dialogs = require('dialogs');
-const dialogs = Dialogs();
-
-const path = require('path');
-const fs = require('fs-extra');
-const os = require('os');
-
-const { PythonShell } = require('python-shell');
-const csvtojson = require("csvtojson");
-
-const preferences = require("../js/preferences.js");
-
-
-// Functions -------------------------------------------------------------------
-/**
- * Converts an Event to a Promise
- * @param  {HTMLElement} item The item to be stuck with an EventListener
- * @param  {string} event The event to be listened for
- * @return {Promise}      Promise that resolves once the event has triggered
- */
-function getPromiseFromEvent(item, event) {
-    return new Promise((resolve) => {
-        const listener = () => {
-            item.removeEventListener(event, listener);
-            resolve();
-        }
-        item.addEventListener(event, listener);
-    })
-}
-
-/**
- * Get click position inside the video element to create a slice
- * @param  {Event} e Event object passed by an EventListener
- */
-function setPtSlice(e) {
-    const video = document.getElementById('video');
-
-    let videoCoords = video.getBoundingClientRect();
-    let xPosition = e.clientX - videoCoords.left;
-    let yPosition = e.clientY - videoCoords.top;
-    if (tempSlicePtStart[0] == -1 && tempSlicePtStart[1] == -1) { // First point
-        tempSlicePtStart = [xPosition / videoCoords.width, yPosition / videoCoords.height];
-        console.log(`First slice pt: ${tempSlicePtStart}`);
-    }
-    else { // Second point
-        tempSlicePtStop = [xPosition / videoCoords.width, yPosition / videoCoords.height];
-
-        // Add new slice to the project
-        let target = findDataByName(vidIdDisp);
-        prj.items[target[0]].items.push(new slice(`slice_${prj.nameNumCount}`, [tempSlicePtStart, tempSlicePtStop], false));
-        prj.nameNumCount++;
-
-        // Set project.saved to false
-        prj.saved = false;
-
-        // Reset
-        console.log(`Second slice pt: ${tempSlicePtStop}`);
-        tempSlicePtStart = [-1, -1];
-        tempSlicePtStop = [-1, -1];
-
-        // Update interface
-        drawSlices();
-        updateAccordions();
-    }
-    // alert('pos x:' + xPosition + '/' + videoCoords.width + ' pos y: ' + yPosition + '/' + videoCoords.height);
-}
-
-/**
- * Draw slices on top of the video canvas
- */
-function drawSlices() {
-    pref = preferences.loadPreferences();
-    const video = document.getElementById('video');
-
-    if (vidIdDisp != '') {
-        let target = findDataByName(vidIdDisp);
-        var videoCoords = video.getBoundingClientRect(); // Access values as height and width
-
-        // Get canvas element
-        let cnvas = document.getElementById('videoCanvas');
-        let ctx = cnvas.getContext('2d');
-
-        // Move/Resize canvas
-        cnvas.style.left = videoCoords.x + 'px';
-        cnvas.width = videoCoords.width;
-        cnvas.height = videoCoords.height;
-
-        console.log('Drawing slices');
-        for (let i = 0; i < prj.items[target[0]].items.length; i++) {
-            // Draw slices
-            ctx.beginPath();
-            ctx.lineWidth = '2';
-            ctx.strokeStyle = pref.sliceColor;
-            ctx.moveTo(prj.items[target[0]].items[i].coord[0][0] * videoCoords.width, prj.items[target[0]].items[i].coord[0][1] * videoCoords.height);
[…]
