+++ frontend.js
@@ -1,0 +1,3 @@
+function changeIcon(newIconPath) {
+    this.src=newIconPath
+}
