### 2023-07-27 #5491 @Kevin-Mattheus-Moerman
@RLado can you please add community/contributing guidelines to your project? You can do this by adding a CONTRIBUTING.md file and to refer to this in your README.md as well. There are many templates online (perhaps this place helps; https://contributing.md/example/) which you might study but you can keep it fairly simple too and simply spell out how you'd like folks to help out/contribute to this project.
