### 2023-07-27 602f5f9
Added contributing guidelines
