### 2023-07-27 2d42edb
Updated README.md with the new contribution guidelines
