### 2023-07-27 #5491 @RLado
@Kevin-Mattheus-Moerman I just added contributing guidelines to the project. They are linked in the README page and [here](https://github.com/RLado/ViMag/blob/master/CONTRIBUTING.md).

Thank you for the suggestion
