### 2023-07-31 125f46a
Release v1.0.0 - Accepted in the Journal of Open Source Software
