+++ README.md
@@ -3,7 +3,7 @@
 
 <a href="https://joss.theoj.org/papers/a0149e43cf19bbafe1cb9eecdcde6189"><img src="https://joss.theoj.org/papers/a0149e43cf19bbafe1cb9eecdcde6189/status.svg"></a>
 
-Vision-based damage detection techinques can reduce sensor deployment costs while providing accurate, useful, and full-field readings of structural behaviour. Our work presents a video processing methodology based on [STB-VMM](https://github.com/RLado/STB-VMM) implemented in a software toolbox that allows the processing of video data to obtain vibrational signatures of complex structures. Our [tests](https://doi.org/10.1016/j.measurement.2022.112218) have demonstrated that this technique enables the identification of very light structural damage in a controlled lab environment. This vision-based technique may not be as precise as other contact-based or laser methods, but on the other hand, it offers an easy to use, effective, full-field, tool for structural health monitoring at a fraction of the cost.
+Vision-based damage detection techniques can reduce sensor deployment costs while providing accurate, useful, and full-field readings of structural behaviour. Our work presents a video processing methodology based on [STB-VMM](https://github.com/RLado/STB-VMM) implemented in a software toolbox that allows the processing of video data to obtain vibrational signatures of complex structures. Our [tests](https://doi.org/10.1016/j.measurement.2022.112218) have demonstrated that this technique enables the identification of very light structural damage in a controlled lab environment. This vision-based technique may not be as precise as other contact-based or laser methods, but on the other hand, it offers an easy-to-use, effective, full-field tool for structural health monitoring at a fraction of the cost.
 
 ## Features
 - Apply state-of-the-art video motion magnification
